Atualização de versão

+++ application/controllers/Vendas.php
@@ -228,6 +228,13 @@ class Vendas extends MY_Controller
         $this->load->model('vendas_model');
 
         $id = $this->input->post('id');
+
+        $editavel = $this->vendas_model->isEditable($id);
+        if (!$editavel) {
+            $this->session->set_flashdata('error', 'Erro ao tentar excluir. Venda já faturada');
+            redirect(site_url('vendas/gerenciar/'));
+        }
+
         $venda = $this->vendas_model->getByIdCobrancas($id);
         if ($venda == null) {
             $venda = $this->vendas_model->getById($id);
@@ -236,7 +243,6 @@ class Vendas extends MY_Controller
                 redirect(site_url('vendas/gerenciar/'));
             }
         }
-
 
         if ($venda->idCobranca != null) {
             if ($venda->status == "canceled") {
@@ -295,6 +301,14 @@ class Vendas extends MY_Controller
         $this->form_validation->set_rules('idProduto', 'Produto', 'trim|required');
         $this->form_validation->set_rules('idVendasProduto', 'Vendas', 'trim|required');
 
+        $editavel = $this->vendas_model->isEditable($this->input->post('idVendasProduto'));
+        if (!$editavel) {
+            return $this->output
+                ->set_content_type('application/json')
+                ->set_status_header(422)
+                ->set_output(json_encode(['result' => false, 'messages' => '<br /><br /> <strong>Motivo:</strong> Venda já faturada']));
+        }
+
         if ($this->form_validation->run() == false) {
             echo json_encode(['result' => false]);
         } else {
@@ -317,6 +331,11 @@ class Vendas extends MY_Controller
                     $this->produtos_model->updateEstoque($produto, $quantidade, '-');
                 }
 
+                $this->db->set('desconto', 0.00);
+                $this->db->set('valor_desconto', 0.00);
+                $this->db->where('idVendas', $this->input->post('idVendasProduto'));
+                $this->db->update('vendas');
+
                 log_info('Adicionou produto a uma venda.');
 
                 echo json_encode(['result' => true]);
@@ -333,6 +352,14 @@ class Vendas extends MY_Controller
             redirect(base_url());
         }
 
+        $editavel = $this->vendas_model->isEditable($this->input->post('idVendas'));
+        if (!$editavel) {
+            return $this->output
+                ->set_content_type('application/json')
+                ->set_status_header(422)
+                ->set_output(json_encode(['result' => false, 'messages' => '<br /><br /> <strong>Motivo:</strong> Venda já faturada']));
+        }
+
         $ID = $this->input->post('idProduto');
         if ($this->vendas_model->delete('itens_de_vendas', 'idItens', $ID) == true) {
             $quantidade = $this->input->post('quantidade');
@@ -344,11 +371,56 @@ class Vendas extends MY_Controller
                 $this->produtos_model->updateEstoque($produto, $quantidade, '+');
             }
 
+            $this->db->set('desconto', 0.00);
+            $this->db->set('valor_desconto', 0.00);
+            $this->db->where('idVendas', $this->input->post('idVendas'));
+            $this->db->update('vendas');
+
             log_info('Removeu produto de uma venda.');
             echo json_encode(['result' => true]);
         } else {
             echo json_encode(['result' => false]);
         }
+    }
+
+    public function adicionarDesconto()
+    {
+        if ($this->input->post('desconto') == "") {
+            return $this->output
+                ->set_content_type('application/json')
+                ->set_status_header(400)
+                ->set_output(json_encode(['messages' => 'Campo desconto vazio']));
+        } else {
+            $idVendas = $this->input->post('idVendas');
+            $data = [
+                'desconto' => $this->input->post('desconto'),
+                'valor_desconto' => $this->input->post('resultado')
+            ];
+            $editavel = $this->vendas_model->isEditable($idVendas);
+            if (!$editavel) {
+                return $this->output
+                    ->set_content_type('application/json')
+                    ->set_status_header(400)
+                    ->set_output(json_encode(['result' => false, 'messages', 'Desconto não pode ser adiciona. Venda não ja Faturada/Cancelada']));
+            }
+            if ($this->vendas_model->edit('vendas', $data, 'idVendas', $idVendas) == true) {
+                log_info('Adicionou um desconto na Venda. ID: ' . $idVendas);
+                return $this->output
+                    ->set_content_type('application/json')
+                    ->set_status_header(201)
+                    ->set_output(json_encode(['result' => true, 'messages' => 'Desconto adicionado com sucesso!']));
+            } else {
+                log_info('Ocorreu um erro ao tentar adiciona desconto a Venda: ' . $idVendas);
+                return $this->output
+                    ->set_content_type('application/json')
+                    ->set_status_header(400)
+                    ->set_output(json_encode(['result' => false, 'messages', 'Ocorreu um erro ao tentar adiciona desconto a Venda.']));
+            }
+        }
+        return $this->output
+            ->set_content_type('application/json')
+            ->set_status_header(400)
+            ->set_output(json_encode(['result' => false, 'messages', 'Ocorreu um erro ao tentar adiciona desconto a OS.']));
     }
 
     public function faturar()
@@ -379,11 +451,13 @@ class Vendas extends MY_Controller
             } catch (Exception $e) {
                 $vencimento = date('Y/m/d');
             }
-
+            $vendas = $this->vendas_model->getById($venda_id);
             $data = [
                 'vendas_id' => $venda_id,
                 'descricao' => set_value('descricao'),
                 'valor' => $this->input->post('valor'),
+                'desconto' => $vendas->desconto,
+                'valor_desconto' => $vendas->valor_desconto,
                 'clientes_id' => $this->input->post('clientes_id'),
                 'data_vencimento' => $vencimento,
                 'data_pagamento' => $recebimento,
